requirements.txt
Django==2.0.4
django-colorfield==0.1.14
django-ratelimit==1.1.0
Khayyam==3.0.17
Pillow==5.1.0
pytz==2018.4
